igdb/pyproject.toml
[tool.poetry]
name = "igdb"
version = "0.1.0"
description = "The IGDB sync service for GameNode."
authors = ["Lamarcke <cassiolamarcksilvafreitas@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
supertokens-python = "^0.16.6"
retry = "^0.9.2"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
